Refactor; use doublestar
